requirements.txt
PyQt5
pybit
python-dotenv
